+++ src/part2-cli/breakpoints/2-dip-toes/index.ts
@@ -71,21 +71,21 @@ interface CLIOptions {
 
 function main(url: string, options?: CLIOptions) {
   return Effect.gen(function* (_) {
-    const headerMap = options?.headers?.reduce((acc, header) => {
+    const headers = options?.headers?.reduce((acc, header) => {
       const [key, value] = header.split(":");
       if (!key || !value) {
         throw new Error("Invalid header");
       }
-      acc.set(key, value);
+      acc.push([key, value]);
       return acc;
-    }, new Map<string, string>());
+    }, new Array<[string, string]>());
 
     const res = yield* _(
       Effect.tryPromise((signal) =>
         fetch(url, {
           ...(options?.method && { method: options.method }),
           ...(options?.data && { body: options.data }),
-          ...(headerMap && { headers: Array.from(headerMap.entries()) }),
+          ...(headers && { headers }),
           signal,
         })
       )