+++ src/part2-cli/breakpoints/5-effect-args-parsing/index.ts
@@ -76,14 +76,22 @@ const CliOptionsLive = Layer.effect(
 const main = Effect.gen(function* (_) {
   const options = yield* _(M.CLIOptions);
 
-  const headerMap = options?.headers?.reduce((acc, header) => {
-    const [key, value] = header.split(":");
-    if (!key || !value) {
-      throw new Error("Invalid header");
-    }
-    acc.set(key, value);
-    return acc;
-  }, new Map<string, string>());
+  const headers = options?.headers
+    ? yield* _(
+        Effect.reduce(
+          options.headers,
+          new Array<[string, string]>(),
+          (acc, header) => {
+            const [key, value] = header.split(":");
+            if (!key || !value) {
+              return Effect.fail(new M.HeaderParseError());
+            }
+            acc.push([key, value]);
+            return Effect.succeed(acc);
+          }
+        )
+      )
+    : [];
 
   const providedFetch = yield* _(M.Fetch);
 
@@ -93,7 +101,7 @@ const main = Effect.gen(function* (_) {
         providedFetch(options.url, {
           ...(options?.method && { method: options.method }),
           ...(options?.data && { body: options.data }),
-          ...(headerMap && { headers: Array.from(headerMap.entries()) }),
+          headers,
           signal,
         }),
       catch: (error) => new M.UnknownError({ error }),